Little fixes

+++ utils/playernav/parse.c
@@ -8,7 +8,7 @@ extern double dumpfreq;
 
 /* Parse command line arguments, of the form host:port */
 int
-parse_args(int argc, char** argv, int* num_bots, char** hostnames, 
+parse_args(int argc, char** argv, size_t* num_bots, char** hostnames, 
            int* ports, double* zoom, int* aa, int* map_idx)
 {
   char* idx;
